# Minor code optim metrics.py