<!--
Thanks for contributing a pull request! Please ensure you have taken a look at
the contribution guidelines: https://scikit-survival.readthedocs.io/en/latest/contributing.html#making-changes-to-the-code
-->

**Checklist**
<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->

- [ ] closes #xxxx : N/A
- [x] pytest passes : _test_metrics.py_ passes
- [ ] tests are included : N/A
- [ ] code is well formatted
- [ ] documentation renders correctly : N/A

**What does this implement/fix? Explain your changes**

It's a minor optimization of the _metrics.py_ file about code writing (mainly factoring) and  performances.
Performance improvements were tested with `timeit`.