@sebp 2023-07-17
Thanks for helping to improve scikit-survival. Could you please provide the sample code that you used for benchmarking too.