@alexandrehuat 2023-07-24
I benchmarked just some code lines with IPython, not whole functions. Here’s a script to reproduce what I did. It doesn’t cover the whole associated commit since not all changes needed benchmark. See the commit. Hope it helps.

Code:
```python
import timeit

import numpy as np


def min_timeit(func, repeat=10, number=10000, globals=None):
    times = timeit.repeat(func, repeat=repeat, number=number, globals=globals)
    return min(times) / number
    
def fmt_timeit(time):
    for p, u in zip([0, 3, 6, 9], " mμn"):
        if time > 10 ** -p:
            stime = time * 10 ** p
            break
    return f"{stime:.3g} {u}s"

def benchmark_commit(line, old_fn, new_fn, repeat=10, number=10000, globals=None):
    print(" Benchmark ".center(32, "="))
    print("Commit line:", line)
    old_time, new_time = [min_timeit(func, repeat, number, globals=globals) for func in [old_fn, new_fn]]
    for tname, time in [("Old", old_time), ("New", new_time)]:
        print(tname + ":", fmt_timeit(time), f"(min exec time, {repeat=}, {number=})")
    time_gain = (new_time - old_time) / old_time
    print(f"Time gain: {time_gain:+.0%}")

def old1():
    ipcw = np.empty(shape, dtype=dtype)
    ipcw[mask] = ipcw_test
    ipcw[~mask] = 0

def new1():
    ipcw = np.zeros(shape, dtype=dtype)
    ipcw[mask] = ipcw_test

def old2():
    np.broadcast_to(np.infty, (1, n_times))

def new2():
    np.full((1, n_times), np.infty)
    

if __name__ == "__main__":
    line = "323-325"
    shape = 1000
    dtype = float
    mask = np.random.randn(shape) > 0
    ipcw_test = np.random.rand(int(mask.sum()))
    benchmark_commit(line, old1, new1, globals=globals())

    line = "477"
    n_times = 1000
    benchmark_commit(line, old2, new2, globals=globals())
``` 

Results:
```
========== Benchmark ===========
Commit line: 323-325
Old: 6.03 μs (min exec time, repeat=10, number=10000)
New: 2.25 μs (min exec time, repeat=10, number=10000)
Time gain: -63%
========== Benchmark ===========
Commit line: 477
Old: 2.26 μs (min exec time, repeat=10, number=10000)
New: 1.09 μs (min exec time, repeat=10, number=10000)
Time gain: -52%
```