@sebp 2024-01-14
Your code doesn't actually benchmark the `cumulative_dynamic_auc` function.

When I do measure `cumulative_dynamic_auc`, I don't see any benefit:

| master branch | PR 392 |
| ----- | ----- |
| 1028944.56 ± 3859.4636526854347 ns | 1029415.92 ± 5412.166604383128 ns |

The code is below.

<details>

```python
import time
import timeit

import pandas as pd
import numpy as np
from sksurv.datasets import load_flchain
from sksurv.linear_model import CoxPHSurvivalAnalysis
from sksurv.preprocessing import OneHotEncoder
from sklearn.preprocessing import StandardScaler
from sksurv.metrics import cumulative_dynamic_auc


if __name__ == "__main__":
    Xin, yin  = load_flchain()
    Xarr = np.tile(Xin.values, (50, 1))
    y = np.tile(yin, 50)
    X = {}
    for i, (n, v) in enumerate(Xin.items()):
        X[n] = pd.Series(Xarr[:, i], name=n, dtype=v.dtype)
    X = pd.DataFrame.from_dict(X)
    print(X.shape)

    Xt = StandardScaler().fit_transform(OneHotEncoder().fit_transform(X).fillna(0))
    m = CoxPHSurvivalAnalysis(alpha=1, ties="efron").fit(Xt, y)
    p = m.predict(Xt)
    t = np.linspace(1, np.quantile(y["futime"], 0.75), 100)

    def score():
        cumulative_dynamic_auc(
            y, y, p, t
        )

    timer = timeit.Timer("score", timer=time.perf_counter_ns, globals=globals())
    r = timer.repeat(repeat=25, number=100000)
    print(np.mean(r), np.std(r))
```

</details>